efw-protocols: add implementation for protocol about hardware control
This commit adds implementation for protocol about hardware control.
The batch of code is moved from internal transaction module to new
module.
