+++ libs/efw/protocols/src/lib.rs
@@ -7,6 +7,7 @@
 //! module.
 
 pub mod flash;
+pub mod hw_ctl;
 pub mod hw_info;
 pub mod transactions;
 pub mod transport;
